Support custom css
This change adds a line to the head template that will look for custom.css
and load this if provided. This is useful for those wanting to extend the
theme without having their custom styles overriden by an update to the base
theme.

Test-information:

Added a custom.css stylesheet in a theme extending x-corporation and
observed that my custom styles were applied.