Improve mobile menu behaviour when page scrolled
If you scroll the page, the mobile menu should probably not get
pushed up by the footer. This makes it awkward to select the first
item in the menu.

This change addresses this by giving the menu container a fixed
position and setting an appropriate z-index.

See issue #31.

Test-information:

Did a quick test on a personal project.

Previously, the mobile menu would get pushed up by the footer.
After this change it is fixed to the height of the viewport.

The menu still scrolls when items overflow.